requirements.txt
# Core Dependencies for Sugarscape Q-Learning with NetLogo Integration
# Tested and working versions for Python 3.13.5

# NetLogo Integration
pynetlogo==0.5.2
JPype1==1.6.0

# Data Science and Analysis
numpy==2.3.2
pandas==2.3.1

# Visualization
matplotlib==3.10.5

# Additional dependencies that may be needed for Q-Learning
# Uncomment if needed for your specific implementation:
# scikit-learn>=1.3.0
# tensorflow>=2.13.0
# torch>=2.0.0
# gym>=0.29.0

# System Requirements:
# - Python 3.13.5
# - Java 23 (JDK/JRE) with JAVA_HOME set
# - NetLogo 6.4.0 (or compatible version)
# 
# Installation Instructions:
# 1. Ensure Java is installed and JAVA_HOME is set
# 2. pip install -r requirements.txt
# 3. Verify NetLogo is installed and accessible